+++ src/main/java/lotto/model/winning/BonusNumber.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.winning;
 
 import java.util.Objects;
 import lotto.exception.ErrorMessage;
